RTK APIs: Extract base query function
