+++ public/app/features/dashboard/api/publicDashboardApi.ts
@@ -1,7 +1,6 @@
-import { BaseQueryFn, createApi } from '@reduxjs/toolkit/query/react';
-import { lastValueFrom } from 'rxjs';
+import { createApi } from '@reduxjs/toolkit/query/react';
 
-import { BackendSrvRequest, config, FetchError, getBackendSrv, isFetchError } from '@grafana/runtime/src';
+import { config, FetchError, isFetchError } from '@grafana/runtime/src';
 import { notifyApp } from 'app/core/actions';
 import { createErrorNotification, createSuccessNotification } from 'app/core/copy/appNotification';
 import { t } from 'app/core/internationalization';
@@ -19,31 +18,11 @@ import {
   PublicDashboardListWithPaginationResponse,
 } from 'app/features/manage-dashboards/types';
 
-type ReqOptions = {
-  manageError?: (err: unknown) => { error: unknown };
-  showErrorAlert?: boolean;
-};
+import { createBaseQuery } from '../../../api/createBaseQuery';
 
 function isFetchBaseQueryError(error: unknown): error is { error: FetchError } {
   return typeof error === 'object' && error != null && 'error' in error;
 }
-
-const backendSrvBaseQuery =
-  ({ baseUrl }: { baseUrl: string }): BaseQueryFn<BackendSrvRequest & ReqOptions> =>
-  async (requestOptions) => {
-    try {
-      const { data: responseData, ...meta } = await lastValueFrom(
-        getBackendSrv().fetch({
-          ...requestOptions,
-          url: baseUrl + requestOptions.url,
-          showErrorAlert: requestOptions.showErrorAlert,
-        })
-      );
-      return { data: responseData, meta };
-    } catch (error) {
-      return requestOptions.manageError ? requestOptions.manageError(error) : { error };
-    }
-  };
 
 const getConfigError = (err: unknown) => ({
   error: isFetchError(err) && err.data.messageId !== 'publicdashboards.notFound' ? err : null,
@@ -51,7 +30,7 @@ const getConfigError = (err: unknown) => ({
 
 export const publicDashboardApi = createApi({
   reducerPath: 'publicDashboardApi',
-  baseQuery: backendSrvBaseQuery({ baseUrl: '/api' }),
+  baseQuery: createBaseQuery({ baseURL: '/api' }),
   tagTypes: ['PublicDashboard', 'AuditTablePublicDashboard', 'UsersWithActiveSessions', 'ActiveUserDashboards'],
   refetchOnMountOrArgChange: true,
   endpoints: (builder) => ({
