+++ public/app/features/migrate-to-cloud/api/baseAPI.ts
@@ -1,38 +1,9 @@
-import { BaseQueryFn, createApi } from '@reduxjs/toolkit/query/react';
-import { lastValueFrom } from 'rxjs';
+import { createApi } from '@reduxjs/toolkit/query/react';
 
-import { BackendSrvRequest, getBackendSrv } from '@grafana/runtime';
-
-interface RequestOptions extends BackendSrvRequest {
-  manageError?: (err: unknown) => { error: unknown };
-  showErrorAlert?: boolean;
-
-  // rtk codegen sets this
-  body?: BackendSrvRequest['data'];
-}
-
-function createBackendSrvBaseQuery({ baseURL }: { baseURL: string }): BaseQueryFn<RequestOptions> {
-  async function backendSrvBaseQuery(requestOptions: RequestOptions) {
-    try {
-      const { data: responseData, ...meta } = await lastValueFrom(
-        getBackendSrv().fetch({
-          ...requestOptions,
-          url: baseURL + requestOptions.url,
-          showErrorAlert: false,
-          data: requestOptions.body,
-        })
-      );
-      return { data: responseData, meta };
-    } catch (error) {
-      return requestOptions.manageError ? requestOptions.manageError(error) : { error };
-    }
-  }
-
-  return backendSrvBaseQuery;
-}
+import { createBaseQuery } from 'app/api/createBaseQuery';
 
 export const baseAPI = createApi({
   reducerPath: 'migrateToCloudGeneratedAPI',
-  baseQuery: createBackendSrvBaseQuery({ baseURL: '/api' }),
+  baseQuery: createBaseQuery({ baseURL: '/api' }),
   endpoints: () => ({}),
 });
