+++ public/app/features/query-library/api/factory.ts
@@ -1,12 +1,14 @@
 import { createApi } from '@reduxjs/toolkit/query/react';
 
-import { baseQuery } from './query';
+import { createBaseQuery } from '../../../api/createBaseQuery';
+
+import { BASE_URL } from './query';
 
 // Currently, we are loading all query templates
 // Organizations can have maximum of 1000 query templates
 export const QUERY_LIBRARY_GET_LIMIT = 1000;
 
 export const queryLibraryApi = createApi({
-  baseQuery,
+  baseQuery: createBaseQuery({ baseURL: BASE_URL }),
   endpoints: () => ({}),
 });
