+++ public/app/features/query-library/api/query.ts
@@ -1,8 +1,3 @@
-import { BaseQueryFn } from '@reduxjs/toolkit/query/react';
-import { lastValueFrom } from 'rxjs';
-
-import { BackendSrvRequest, getBackendSrv, isFetchError } from '@grafana/runtime/src/services/backendSrv';
-
 import { getAPINamespace } from '../../../api/utils';
 
 /**
@@ -23,29 +18,3 @@ export enum QueryTemplateKinds {
  * @alpha
  */
 export const BASE_URL = `/apis/${API_VERSION}/namespaces/${getAPINamespace()}`;
-
-interface QueryLibraryBackendRequest extends BackendSrvRequest {
-  body?: BackendSrvRequest['data'];
-}
-
-export const baseQuery: BaseQueryFn<QueryLibraryBackendRequest, unknown, Error> = async (requestOptions) => {
-  try {
-    const responseObservable = getBackendSrv().fetch({
-      url: `${BASE_URL}/${requestOptions.url ?? ''}`,
-      showErrorAlert: true,
-      method: requestOptions.method || 'GET',
-      data: requestOptions.body,
-      params: requestOptions.params,
-      headers: { ...requestOptions.headers },
-    });
-    return await lastValueFrom(responseObservable);
-  } catch (error) {
-    if (isFetchError(error)) {
-      return { error: new Error(error.data.message) };
-    } else if (error instanceof Error) {
-      return { error };
-    } else {
-      return { error: new Error('Unknown error') };
-    }
-  }
-};
